requirements.txt
arxiv>=2.1.0
requests>=2.31.0
beautifulsoup4>=4.12.0
PyYAML>=6.0.0
pdf2image>=1.17.0
pytest>=8.0.0
